+++ src/interfaces/StaticMessageInterface.php
@@ -51,7 +51,6 @@ class StaticMessageInterface
             'language' => Type::string(),
             'category' => Type::string(),
             'dateCreated' => Type::string(),
-            'dateUpdated' => Type::string(),
         ];
     }
 }